Make QuadraticSieveScheduler compile on 32bit systems

+++ src/elves/quadratic_sieve/QuadraticSieveScheduler.cpp
@@ -193,7 +193,7 @@ vector<double> QuadraticSieveScheduler::determineChunkSizes(const BigInt& start,
 
     uint64_t intervalLength = intervalLengthBigInt.get_ui();
     cout << "IntervalLength: " << intervalLength << endl;
-    double basicChunkSize = div_ceil(intervalLength, communicator.size());
+    double basicChunkSize = div_ceil(intervalLength, static_cast<uint64_t>(communicator.size()));
     auto weights = communicator.weights();
     cout << "Weights: " << weights << endl;
